+++ tests/integration/test_cli.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 
 import yaml
 import pytest
-from tests.integration.conftest import contains_expected_workflows
+from integration.conftest import contains_expected_workflows
 from databricks.labs.dqx.cli import (
     open_remote_config,
     installations,
